+++ CHANGELOG.md
@@ -2,6 +2,61 @@
 
 All notable changes will be documented in this file.
 
+#### [0.6.4] - 2020-03-06
+
+-   Update to support multicollateral contract changes
+
+-   `Positions` added methods
+-   `createInThreeActions`
+-   `getPositionByMaker`
+-   `getLatestUserPosition`
+-   `getLtvRatiosTable`
+-   `getPositionLtvRatio`
+
+-   `Liquidator` added methods
+-   `getSettings`
+-   `getNutCollatBalance`
+
+-   `Liquidator` updated methods
+-   `marginCallAndBuyoutCollat` instead of `marginCallAndBuyoutEos`
+-   `getCollatBalance` instead of `getEosBalance`
+-   `transferEos` removed
+
+-   `Governance` added methods
+-   `getParameters`
+-   `getVotesForAccount`
+
+-   New `Positions` interfaces
+-   `TokenRate`
+-   `LtvRatios`
+
+-   New `Positions` settings in interface `EosdtContractSettings` 
+-   `collateral_account`
+-   `collateral_token`
+
+-   Updated `Positions` interfaces
+-   `TokenRate_deprecated` instead of old `TokenRate`
+
+-   New `Liquidator` interface `LiquidatorSettings` 
+
+-   Updated `Liquidator` parameters in interface `LiquidatorParameters`
+-   `collat_balance` instead of old `eos_balance`
+
+-   New `Governance` interfaces
+-   `GovernanceParameters`
+
+-   Updated `Governance` settings in interface `GovernanceSettings`
+-   `position_account` instead of old `eosdtcntract_account`
+
+#### [0.6.38] - 2020-03-04
+
+-   Updated `Positions` following methods due to contracts update
+
+-   `getContractEosAmount` become deprecated
+-   `getContractTokenAmount` instead of `getContractEosAmount`
+-   `getRates` become deprecated
+-   `getRelativeRates` instead of `getRates` updated to work with new table `oraclerates`
+
 #### [0.6.37] - 2019-11-20
 
 -   Added `BpManager` with following methods to help manage block producers positions